Simplify DragAndDrop into a single file
